Boot order change for k8s cluster
It seems that k8s will load all the required pods as resources are
available. If beefier machines join, resources are not re-scheduled to
better use the resources. So it's possible small containers will first
be scheduled in the beefier machines and then heavy containers are not
schedulable. This is an attempt to make the cluster first use the small
machines (and thus only schedule small containers) and then enable the
beefier machines to run the actual heavy resources. This seems a bit
like the slow joiner syndrome.